Reducing repeated code in tests
