+++ test/test_socket.py
@@ -7,8 +7,11 @@ from config import TESTDB, TESTDB_PATH, TestConfig
 from worker import conn
 from redis import ConnectionError
 import random
+from pprint import pprint
+# from .conftest import mock_players, PLAYERS_TO_GEN, is_redis_available
 
 
+PLAYERS_TO_GEN = 5
 """
 Must have redis instance running before this test is invoked
 """
@@ -22,62 +25,62 @@ def is_redis_available(conn):
 pytestmark = pytest.mark.skipif(not is_redis_available(conn), reason="redis instance is not available")
 
 
-PLAYERS_TO_GEN = 5
-mock_players = {}
+# PLAYERS_TO_GEN = 5
+# mock_players = {}
 
-@pytest.fixture(scope='session')
-def app(request):
-    _app.config.from_object(TestConfig)
-    ctx = _app.app_context()
-    ctx.push()
-    def teardown():
-        ctx.pop()
-    request.addfinalizer(teardown)
-    return _app
+# @pytest.fixture(scope='session')
+# def app(request):
+#     _app.config.from_object(TestConfig)
+#     ctx = _app.app_context()
+#     ctx.push()
+#     def teardown():
+#         ctx.pop()
+#     request.addfinalizer(teardown)
+#     return _app
 
-@pytest.fixture(scope='session')
-def db(app, request):
-    if os.path.exists(TESTDB_PATH):
-        os.unlink(TESTDB_PATH)
-    def teardown():
-        _db.drop_all()
-        os.unlink(TESTDB_PATH)
-        conn.shutdown(nosave=True)
-    _db.init_app(app)
-    _db.create_all()
-    configure(app, _db, request)
-    print(_db.engine.table_names())
-    request.addfinalizer(teardown)
-    return _db
+# @pytest.fixture(scope='session')
+# def db(app, request):
+#     if os.path.exists(TESTDB_PATH):
+#         os.unlink(TESTDB_PATH)
+#     def teardown():
+#         _db.drop_all()
+#         os.unlink(TESTDB_PATH)
+#         conn.shutdown(nosave=True)
+#     _db.init_app(app)
+#     _db.create_all()
+#     configure(app, _db, request)
+#     print(_db.engine.table_names())
+#     request.addfinalizer(teardown)
+#     return _db
 
-def configure(app, db, request):
-    with app.test_client() as client:
-        global mock_players
-        for i in range(1, PLAYERS_TO_GEN+1):
-            models.Player.create(username='Randomplayer{}'.format(i), 
-                                 email='Rando{}@somewhere.somehow'.format(i), 
-                                 password='Rand{}'.format(i))
-            # response = client.post(
-            #     '/api/player/create',
-            #     data=json.dumps({'username' : 'Randomplayer{}'.format(i),
-            #                     'email' : 'Rando{}@somewhere.somehow'.format(i),
-            #                     'password': 'Rand{}'.format(i)}),
-            #     content_type='application/json'
-            # )
-        mock_players = {player.id : player for i, player in enumerate(models.Player.query.all())}
-        models.Lobby.create(game='DOTA 2', 
-                            short_description='Some MMORPG game', 
-                            cap=8)
+# def configure(app, db, request):
+#     with app.test_client() as client:
+#         global mock_players
+#         for i in range(1, PLAYERS_TO_GEN+1):
+#             models.Player.create(username='Randomplayer{}'.format(i), 
+#                                  email='Rando{}@somewhere.somehow'.format(i), 
+#                                  password='Rand{}'.format(i))
+#             # response = client.post(
+#             #     '/api/player/create',
+#             #     data=json.dumps({'username' : 'Randomplayer{}'.format(i),
+#             #                     'email' : 'Rando{}@somewhere.somehow'.format(i),
+#             #                     'password': 'Rand{}'.format(i)}),
+#             #     content_type='application/json'
+#             # )
+#         mock_players = {player.id : player for i, player in enumerate(models.Player.query.all())}
+#         models.Lobby.create(game='DOTA 2', 
+#                             short_description='Some MMORPG game', 
+#                             cap=8)
 
-@pytest.fixture
-def client(db, app, request):
-    # db_fd, app.config['DATABASE'] = tempfile.mkstemp()
-    # app.config['TESTING'] = True
-    # with app.socket_client() as client:
-    #     yield client
-    print(models.Player.query.all())
-    with app.test_client() as client:
-        return client
+# @pytest.fixture
+# def client(db, app, request):
+#     # db_fd, app.config['DATABASE'] = tempfile.mkstemp()
+#     # app.config['TESTING'] = True
+#     # with app.socket_client() as client:
+#     #     yield client
+#     print(models.Player.query.all())
+#     with app.test_client() as client:
+#         return client
 
 def login(player, app_client):
     return app_client.post(
@@ -100,7 +103,7 @@ def logout(app_client):
 #     #     yield client
 #     return mock_players
 
-def test_connection(app, client):
+def test_connection(app, client, mock_players):
     namespace = '/connection'
     response = login(mock_players[1], client)
     print(response.get_data(as_text=True))
@@ -110,15 +113,15 @@ def test_connection(app, client):
     received = socket_client.get_received(namespace)
     socket_client_attack.connect(namespace=namespace, query_string='player=1')
     # assert not socket_client_attack.is_connected(namespace)
+    print("Connection status returns")
     print(received)
     socket_client.disconnect(namespace=namespace)
     assert len(received) == 1
     assert received[0]['args'][0]['status'] == 'Connected'
     assert not socket_client.is_connected(namespace)
 
-def test_players_match(app):
+def test_players_match(app, mock_players):
     namespace = '/connection'
-    global mock_players
     clients = {}
     socket_clients = {}
     for player in mock_players.values():
@@ -153,9 +156,8 @@ def test_players_match(app):
         logout(client)
         assert not socket_client.is_connected(namespace)
 
-def test_room_list(app, client):
+def test_room_list(app, client, mock_players):
     namespace = '/connection'
-    global mock_players
     clients = {}
     socket_clients = {}
     for player in mock_players.values():
@@ -174,7 +176,15 @@ def test_room_list(app, client):
         }),
         content_type='application/json'
     )
-    print(json.loads(response.get_data(as_text=True)))
+
+    data = json.loads(response.get_data(as_text=True))
+    assert 0 <= len(data) <= len(mock_players)
+    pprint('Returned data on list rooms')
+    pprint(data)
+    for socket_client, client in zip(socket_clients.values(), clients.values()):
+        socket_client.disconnect(namespace=namespace)
+        logout(client)
+        assert not socket_client.is_connected(namespace)
